Squelch command output in console output.
The output of individual commands is incredibly spammy when running
parallel execution.  The full output is still logged to file, but the
default console view just shows which build commands are being executed
(and errors if something goes wrong).  To bring back the full spew, `-v`
or `--verbose` can be flagged on the command line.