Make hostlist processing not rely on hostnames being unique
The hostlist processing currently works still on pure hostnames as a
legacy from before we made proper host objects to pass around. This
works fine while all hostnames are unique as we just map to and from
host objects using the name.

With the hippo inventory source, hostnames can be duplicated if the name
tags on the instances are configured in such a way. This happens in prod
with the "app-pending" instances during kick. This is probably not
actually a problem for the pending boxes as they're still being
provisioned and will catch up later, but it's good to just eliminate
this subtlety before it becomes an issue.