Lazily resolve host aliases to avoid blocking on unrelated projects
When a service is being spun up or down we'll have aliases in the
rollingpin config that don't resolve to any live servers. Previously,
this would cause rollout to complain about the alias being unresolvable
since it was trying to prevent silent failure. That is pretty
frustrating for people trying to deploy to unrelated projects. This
makes it so that aliases are only resolved when specified on the command
line and so we'll only complain at you if we're unable to resolve hosts
for something you specifically requested.