Support multiple processes per SHA per host.
There was a problem with the way we were doing this before: it made the
assumption we only had one version of each running SHA on each host.
This way of structuring the coverage report will allow apps to specify
the total number of processes running a SHA, which we can then aggregate
across all hosts.