Don't allow rounding to unleash deploys
When specifying a percentage to progress to in deploy that was less than
the jump to the next host, rounding would make the pause_after setting
go to zero which indicated to the deployer that it was time to unleash
the deploy. This is very surprising and dangerous to the user!