Profile support for rollingpin
Attempt to fulfill the desire for rollingpin to be aware and
contextually configurable for multiple projects.  Before the internal
configuration was global and the same across different project deploys.
This adds a required argument that defines a "profile" which ends up
being an additional configuration file to be loaded and can provide more
defaults to dictate a sane default behavior based on a profile.  Because
of how the configuration loading is done, *any* config value that works
in the global config can be overriden via the profile configuration.

For example, `$ rollout -h apps -d public -r all` could become `$
rollout public` where the `public` profile would be an ini file that
sets defaults for the three argumented flags.

Currently (like the location of the global and user scoped config files)
the profiles are hardcoded to live in `/etc/rollingpin.d/` but can be
overriden with optional kwargs for the various related functions.  Three
additional configuration options have been added for setting default
values for the `-h`, `-c`, and `-r` args, more can be added.