Clean up Action functionality
After some research, it appears that the `argparse` library is setting
the default as the same ref (because assignment with lists is by
reference) so this now does a test to check if the namespace value is
the same ref as the default, if it is, it does a shallow copy of it and
then acts on it (and stores it on the namespace, breaking the double ref
issue).

I may spend some time later trying to reproduce this in a simple way and
possibly file a bug with the cpython project, but ultimately this
functionality is *very* not well documented across the various usecases
possible.