Prompt users with warnings if deploying in unsafe conditions
This checks against a new endpoint in Harold to see if it's currently
outside of sanctioned deploy times or other deploys are in progress and
displays a warning prompt to the user if either condition applies. This
is intended to give a little extra indication that extra consideration
should be applied and not a hard rule.