PR Feedback: restart ordering
Talked out of band with Neil and he raised a concern around the ordering
of commands with the `restart` special case.  After some discussion, we
are opting to remove the arbitrary ordering for restarts and instead
keeping the restart action to the end of the commands.  This means that
the parser will store the value separately and then append it after the
parsing is done (so as to keep change throughout the code minimal).