PR Feedback: rework new restart argument logic
After some discussion, it was raised (and agreed) that the cleanup in
the `_main` function was messy.  After some talk, it was felt like
exposing the entire `ArgumentParser` was heavy-handed and I have
attempted to provide just an encapsulating `parse_args` function which
performs the construction, setup, base parsing, and post parsing cleanup
(i.e. the `restart` translation into `commands`).  This obfuscates the
internal attachment to the `argparse` library outside of the module and
just treats the parsing as a single first order function rather than an
object oriented construction and (single) method caller.