Partially deprecate `no-harold`
Now that profiles are in use and notifying the correct harold backend,
the `no-harold` flag is not as necessary, so it is being "deprecated"
and will produce a message.  There is a new flag called
`really-no-harold` that will behave the same way.  At some point we may
switch back to the `no-harold` form but after people get used to the new
behavior.