Encode a safe and reasonable deploy process in user prompts
The previous UI was serviceable if you knew what you were doing but left
a lot to the imagination. Newer engineers didn't necessarily know what
the proper procedure was and so we've had some issues with inconsistent
deploy policies. This new UI gives the user a more guided path with
prompts that try to nudge in the direction of our best-practice deploys.
This is meant only to help and not be a panacea in itself as training,
education, and social practices are the core problem here.